Aj/droan debug (#163)
* Fix check for ros_ws install folder

* Start refactoring/cleaning/organizing this weird code

* Rename MapRepresentation-->CostMapInterface; rename DisparityMapRepresentation-->DisparityGraphCostMap

* Replace ifndefs with pragma once

* Rename .h to .hpp

* Refactor droan

Refactor to use concept of safety cost

* Fix loop structure

* Change for loop int to size_t

* Fix compile issue

* Suppress debugs logs

* Fix run time crash

* Parameter tuning

* Better debugging

* Set output screen for cout

* Decrease random walk angle so that drone can fly forward

* Use .at() syntax instead of [] to prevent segfault

* Tune parameters; match zedx, apply 1 hack for expansion

* Rename robot_radius to expansion_radius to be more accurate

* Turn array into vector and remove hardcoded value

* Set max velocity to 20.

* Increase lookahead time to increase speed

* Update rviz

+++ common/ros_packages/airstack_common/src/vislib.cpp
@@ -124,7 +124,7 @@ namespace vis{
     va_list list;
     va_start(list, lines);
 
-    for(int i = 0; i < lines; i++){
+    for(size_t i = 0; i < lines; i++){
       float x1 = va_arg(list, double);
       float y1 = va_arg(list, double);
       float z1 = va_arg(list, double);
@@ -148,7 +148,7 @@ namespace vis{
     Marker marker(frame_id, stamp, ns, id++, visualization_msgs::msg::Marker::LINE_STRIP);
     marker.set_scale(width, 0, 0);
     
-    for(int i = 0; i < points.size(); i++){
+    for(size_t i = 0; i < points.size(); i++){
       PointColor& pc = points[i];
       marker.add_point(pc.x, pc.y, pc.z);
       marker.add_color(pc.r, pc.g, pc.b, pc.a);
@@ -177,7 +177,7 @@ namespace vis{
   }
 
   void MarkerArray::overwrite(){
-    for(int i = 0; i < markers.size(); i++)
+    for(size_t i = 0; i < markers.size(); i++)
       markers[i].set_action(visualization_msgs::msg::Marker::DELETE);
     id = 0;
   }
@@ -198,7 +198,7 @@ namespace vis{
   visualization_msgs::msg::MarkerArray MarkerArray::get_marker_array(){
     visualization_msgs::msg::MarkerArray marker_array;
     
-    for(int i = 0; i < markers.size(); i++){
+    for(size_t i = 0; i < markers.size(); i++){
       marker_array.markers.push_back(markers[i].get_marker());
     }
